Fix for CharmNt SVS_00017963 and  SVS_00018074 (Please see descriptions below)
SVS_00017963:On Configure/JVM Page if you type "JVM" into the search textbox many more results are displayed than expected (zero)

SVS_00018074: When searching for "9" on groups page group 25 is returned as well. it could be that some attributes about the items searched for have the number 9 associated with them but it is not an item that is displayed in the fields.. example search for 9 in Group, notice group test 25 is also returned, navigate to JVM screen, search for 25 and notice that https for group test 25 item has "89"

This is a shortcoming of jQuery Datatable version 1.9.4. Please see http://datatables.net/forums/discussion/20831/exclude-html-contents-classes-from-text-search.

The said problem is also seen in group, web server and web application configuration pages.