Fix potential  problem wherein web server generation can be reported as successful even though it isn't because of the finally clause
Details:

A runtime exception that is not caught will make the code go to the finally clause with the didSucceed being true  if the said variable is true by default. To correct this, the default value of didSucceed should be false and later set to true at the end of the the "happy path" code block.